backend/requirements.txt
fastapi
uvicorn
anthropic
python-multipart
pydantic
python-dotenv
requests
beautifulsoup4
newspaper3k
scikit-learn
pandas
numpy
nltk
joblib
Pillow
pytesseract
transformers
torch
reportlab
tldextract
python-whois
websockets
